+++ property_core/__init__.py
@@ -9,7 +9,13 @@ from property_core.block_service import analyze_blocks
 from property_core.companies_house_client import CompaniesHouseClient
 from property_core.epc_client import EPCClient
 from property_core.enrichment import compute_enriched_stats, enrich_comps_with_epc
+from property_core.models.block import BlockAnalysisResponse, BlockBuilding
+from property_core.models.companies_house import CompanyRecord, CompanySearchResult
+from property_core.models.epc import EPCData
 from property_core.models.postcode import PostcodeResult
+from property_core.models.ppd import PPDCompsResponse, PPDTransaction, PPDTransactionRecord
+from property_core.models.report import PropertyReport, RentalAnalysis, YieldAnalysis
+from property_core.models.rightmove import RightmoveListing, RightmoveListingDetail
 from property_core.planning_service import PlanningService
 from property_core.postcode_client import PostcodeClient
 from property_core.ppd_client import PricePaidDataClient
@@ -22,16 +28,16 @@ from property_core.rightmove_location import RightmoveLocationAPI
 from property_core.rightmove_scraper import fetch_listing, fetch_listings
 
 __all__ = [
+    # Services
     "CompaniesHouseClient",
     "EPCClient",
     "PlanningService",
     "PostcodeClient",
-    "PostcodeResult",
     "PPDService",
     "PricePaidDataClient",
     "PropertyReportService",
     "RightmoveLocationAPI",
-    "StampDutyResult",
+    # Functions
     "analyze_blocks",
     "analyze_rentals",
     "calculate_stamp_duty",
@@ -41,4 +47,20 @@ __all__ = [
     "fetch_listing",
     "fetch_listings",
     "match_epc_address",
+    # Models
+    "BlockAnalysisResponse",
+    "BlockBuilding",
+    "CompanyRecord",
+    "CompanySearchResult",
+    "EPCData",
+    "PostcodeResult",
+    "PPDCompsResponse",
+    "PPDTransaction",
+    "PPDTransactionRecord",
+    "PropertyReport",
+    "RentalAnalysis",
+    "RightmoveListing",
+    "RightmoveListingDetail",
+    "StampDutyResult",
+    "YieldAnalysis",
 ]